Add a critical section to Reservation
Added a critical section which contains some of the reservation
validations and the possible reservation saving. It is implemented by
selecting the resource in question for update, so that it will block
subsequent requests which have the same resource. This is needed so
that concurrent requests cannot be validated incorrectly. Previously it
was possible that multiple requests passed the validation before any of
them was saved.

For the critical section to work it needs to be in a transaction,
so enabled ATOMIC_REQUESTS, which seems to be a good thing for other
stuff as well, atleast for now.