Return reservations in .xlsx format
Reservation list and detail endpoints return Excel .xlsx files if
the request's Accept header is to

application/vnd.openxmlformats-officedocument.spreadsheetml.sheet

The .xlsx files contain the same reservations as the endpoints normally
return with fields unit (name), resource (name), begin and end times,
user (email) and comments. User and comments are only added if the
requester has the rights to see those.

Internally the implementation uses the same serializer as before which
makes things a bit more complicated, but it also makes it possible to
use the same field restriction logic for .xlsx files.