Add Reservation access code to API
The access code can only be set if it's enabled in the Resource, and
it cannot be changed when modifying a Reservation.

The access_code field is excluded from returned fields if the current user
isn't a staff member or have can_view_reservation_access_code permission
on the Unit.

Refs #147