Allow setting Units and Resources reservable only for X days in advance
Added field reservable_beforehand_in_days field to Unit and Resource,
which can be used to set the amount of days the entity can be reserved
in advance. If a Resource does not have a value, the Unit's value is
used as a fallback.

In API there is also reservable_before field which determines the
exact datetime the entity can be reserved until.

Freezegun lib is added to keep tests deterministic.

Closes #156