Refactor commenting a bit
Changed the way commentable models are declared from a single string
to a tuple containing an API name and the actual model. This way the
name used in the API can be chosen independent from the model's content
type name, and the whole implementation is also more robust.

Visible changes:
  * Catering order comments are now referred with target_type
    catering_order instead of cateringorder
  * Other than commentable models' comments are excluded in the API
    (of course normally those should not exists in the first place)