Implement period to resource available between filter
Added an optional third parameter to resource available between
filter, which if given is used to determine a minimum free slot length
that must exists in the main time range. Start and end times must be on
the same day.

Known issue: because of limits in the current opening time implementation,
it is not possible to have end time at midnight, so if there is a free
time slot just before midnight, it cannot be detected.