Move styles from resourceForm.css to main.css
The main CSS file was accidentally renamed to resourceForm.css in commit
a8c413b44.  Fix this by introducing a new bundle for the styles and
removing the style import from the resourceFormIndex.js.

Also add a "ls -l" for the main.css file to the Travis script to make
sure that the main.css got generated.  This works two ways: It will fail
the build if it's not there, because ls will exit with non-zero status
and it will show the timestamp and size of the file in the build log
when it got build correctly.