Add API endpoint for checking Order prices
Because one probably wants to check an order's price info before actually
creating the order, added an API endpoint for that. It is used by POSTing
order data to /v1/order/check_price/. The data should be the same that is
used to create the order, with the exception that "return_url" isn't needed.