respa_exchange: Implement heuristics for determining the real reservation organizer name
Sometimes Exchange sets the Organizer element to point to the resource mailbox.
In these cases, we need to resort to heuristics to figure out who really
made the reservation.

If the event has required attendees, we choose the first attendee as the host.
If not, we use the DisplayTo field as the reserver name.