Make resource serializer selection more straightforward
Do not allow to select resource serializers on settings but instead select them
in the code based on whether payments are enabled or not. While at beginning it
seemed to solve the resource serializer selection issue quite nicely, in the
end the settings based solution didn't really add anything else than more
complexity.