Expire too old reservations without orders
Previously the auto-expiration management command expired orders which in turn
set their reservations cancelled. However, if there was a reservation for which
an order is never created (should not normally happen), that reservation didn't
get cancelled ever. Now reservations that are waiting for an order will be set
cancelled if an order isn't added in 30 seconds.

This commit includes also some name changes to already existing things.
Naming on this case turned out to be even more difficult than it is normally,
and the results pretty much suck as well.