Change the way paid reservations are created
Previously in order to create a reservation with an order, one needed
to first create the reservation, and then the order on a separate
request. From now on, orders are created using field "order" inside
reservation data when POSTing a new reservation. The "order" field
must contain all the same fields as before. The reservation endpoint
is only used to create the orders; when updating a reservation, the
"order" field is read-only.

In addition to the functional changes, carried out some refactoring
related to order / reservation serializers and importing of those,
the biggest change being that payments' api has been divided into
multiple modules.