Remove tax amount and pretax price calculations
Now that we are using after-tax prices for products, it actually seems like we
don't really need pre-tax prices or tax amounts at all (exception being product
pretax price because Bambora Payform requires that). Because those
calculations can be implemented in multiple ways (the results vary depending
on when you do rounding), and it is pretty unclear how exactly we need to do
them, let's just not. We can bring these easily back later when/if needed.